requirements.txt
fastapi>=0.115.6,<0.120
uvicorn[standard]>=0.34.0,<1.0
pydantic>=2.11.0,<3.0
pydantic-settings>=2.7.0,<3.0
email-validator>=2.2.0
SQLAlchemy>=2.0.36,<3.0
psycopg[binary]>=3.2.4
alembic>=1.14.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.20
slowapi>=0.1.9
qrcode[pil]>=8.0
segno>=1.6.1
